Update 'web_cart' migration to change data types for 'id' and 'user_id' fields
- Modified the 'id' field from bigint to int to align with application requirements.
- Changed the 'user_id' field from bigint to int, ensuring consistency in data types across the migration.
- These adjustments improve compatibility and maintainability of the database schema for cart functionality.

+++ web/migrations/0008_update_cart_session_key.py
@@ -26,11 +26,11 @@ def check_and_create_cart_table(apps, schema_editor):
                 cursor.execute(
                     """
                     CREATE TABLE web_cart (
-                        id bigint NOT NULL AUTO_INCREMENT PRIMARY KEY,
+                        id int NOT NULL AUTO_INCREMENT PRIMARY KEY,
                         session_key varchar(40) NOT NULL DEFAULT '',
                         created_at datetime(6) NOT NULL DEFAULT CURRENT_TIMESTAMP(6),
                         updated_at datetime(6) NOT NULL DEFAULT CURRENT_TIMESTAMP(6) ON UPDATE CURRENT_TIMESTAMP(6),
-                        user_id bigint NULL,
+                        user_id int NULL,
                         CONSTRAINT fk_cart_user
                             FOREIGN KEY (user_id)
                             REFERENCES auth_user(id)
